Move the Temporal equality tests to a .test.ts file
deep-equals.spec.ts returns to its original contents; the Temporal
coverage lives in its own dedicated test file beside it.
